backend/backend-service-for-recommendation-after-prediction/requirements.txt
google-genai
flask
flask-cors